consolidate email icon
